requirements.txt
pandas==2.2.2
scikit-learn==1.5.1
xgboost==2.1.1
pyyaml==6.0.2
joblib==1.4.2
